converted to using npm and browserify
the idea is that only one js file needs to be loaded -- all dependencies (codemirror, magnific-popup, etc) as well as the logic of the webpage are included in one bundle.js file
