--- data/intro.js
@@ -1,3 +1,0 @@
-/** Load before other data files **/
-const exercises = [];
-const solutions = {};
