+++ exercise.html
@@ -3,14 +3,14 @@
   <head>
     <meta charset="utf-8">
     <title>CodingJS</title>
-    <link rel="stylesheet" href="css/style.css" media="screen" title="no title">
     <link href="https://fonts.googleapis.com/css?family=Trirong" rel="stylesheet">
+    <link rel="stylesheet" type="text/css" href="css/codemirror.css">
+    <link rel="stylesheet" href="css/style.css" media="screen">
     <link rel="stylesheet" href="css/magnific-popup.css">
-    <link rel="stylesheet" type="text/css" href="codemirror/codemirror.css">
   </head>
   <body>
     <div class="main">
-      <p class="heading"><a class="heading" href="."><img src="img/favicon-96x96.png" width="32px" alt="favicon" />&nbsp;<b>CodingJS</a></b></p>
+      <p class="heading"><a class="heading" href="."><img src="img/favicon-96x96.png" width="32px" />&nbsp;<b>CodingJS</a></b></p>
       <p class="subheading">Solve programming problems in Javascript!</p>
 
       <a class="links" id="previous" href="#">previous</a>&nbsp;~&nbsp;<a class="links" id="next" href="#">next</a>&nbsp;~&nbsp;<a href="jsref.html" class="links ajax-popup-link" id="jsref-link">JS Syntax Help</a><br>
@@ -60,35 +60,6 @@
   </div>
 
 </body>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.35.0/codemirror.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.35.0/addon/comment/comment.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.35.0/addon/edit/matchbrackets.min.js"></script>
-
-
-  <script src="codemirror/codemirror-jsmode.js"></script>
-  <script type="text/javascript">
-    var editor = CodeMirror.fromTextArea(document.getElementById("answer"), {
-      lineNumbers: true,
-      matchBrackets: true,
-      mode: "text/javascript"
-    });
-  </script>
-
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.1/jquery.js" charset="utf-8"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/lodash.js/4.16.6/lodash.js" charset="utf-8"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/FileSaver.js/1.3.3/FileSaver.min.js"></script>
-  <script src="js/popup.js"></script>
-
-  <script src="data/intro.js"></script>
-  <script src="data/warmup.js"></script>
-  <script src="data/string.js"></script>
-  <script src="data/array.js"></script>
-  <script src="data/logic.js"></script>
-  <script src="data/ap.js"></script>
-  <script src="data/recursion.js"></script>
-  <script src="data/quiz.js"></script>
-  <script src="helpers.js"></script>
-  <script src="main.js"></script>
-
+<script src="dist/bundle.js"></script>
 
 </html>
