+++ index.html
@@ -3,13 +3,13 @@
 <head>
   <meta charset="utf-8">
   <title>CodingJS</title>
-  <link rel="stylesheet" href="css/style.css" media="screen" title="no title">
   <link href="https://fonts.googleapis.com/css?family=Trirong" rel="stylesheet">
-  <link rel="stylesheet" type="text/css" href="codemirror/codemirror.css">
+  <link rel="stylesheet" href="css/style.css" media="screen">
+  <link rel="stylesheet" type="text/css" href="css/codemirror.css">
 </head>
 <body>
   <div class="main">
-    <p class="heading"><img src="img/favicon-96x96.png" width="32px" alt="favicon" />&nbsp;<b>CodingJS</b></p>
+    <p class="heading"><img src="img/favicon-96x96.png" width="32px" />&nbsp;<b>CodingJS</b></p>
     <p class="subheading">Solve programming problems in Javascript!<br/>
        <i>This site will only store your progress within your current web browser.<br>
           To keep your solutions permanently, be sure to save your solution as a file.</i>
@@ -26,24 +26,9 @@
 
     </div>
 
-    <!-- <a href=`exercise?name=scoresIncreasing&title=AP-1`>scoresIncreasing</a>
-    <a href=`exercise?name=scores100&title=AP-1`>scores100</a>
 
-    <a href=`exercise?name=diff21&title=warm-up1`>diff21</a> -->
-    <footer><hr><i>Adapted by Dan Schellenberg from a JavaScript version made by the-winter, based on Nick Parlante's <a class="links"href="http://codingbat.com/" target="_blank">CodingBat</a>.</i></hr></footer>
+    <footer><hr><i>Adapted by Dan Schellenberg from a JavaScript version made by the-winter, based on Nick Parlante's <a class="links"href="http://codingbat.com/">CodingBat</a>.</i></hr></footer>
   </body>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.1/jquery.js" charset="utf-8"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/lodash.js/4.16.6/lodash.js" charset="utf-8"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.35.0/codemirror.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/FileSaver.js/1.3.3/FileSaver.min.js"></script>
-  <script src="data/intro.js"></script>
-  <script src="data/warmup.js"></script>
-  <script src="data/string.js"></script>
-  <script src="data/array.js"></script>
-  <script src="data/logic.js"></script>
-  <script src="data/ap.js"></script>
-  <script src="data/recursion.js"></script>
-  <script src="helpers.js"></script>
-  <script src="glossary.js"></script>
+  <script src="dist/bundle.js"></script>
 
 </html>
