+++ src/data/recursion.js
@@ -1,4 +1,4 @@
-exercises.push(
+module.exports = [
   { question: 'Given n of 1 or more, return the factorial of n, which is n * (n-1) * (n-2) ... 1. Compute the result recursively (without loops).',
     title: 'Recursion-1',
     name: 'factorial',
@@ -475,356 +475,4 @@ exercises.push(
   "('hiHellohihihi', 'o')",
   "('hiHellohihihi', 'll')"
 ] }
-)
-
-/** --- solutions --- **/
-
-solutions.factorial =  function (n) {
-   if (n === 0) {
-       return 1;
-   }
-    return n * factorial(n-1)
-}
-
-solutions.bunnyEars = function (bunnies) {
-  let bunnyEarsRecursive = function(bunnies){
-    if (bunnies <= 0) return 0;
-    let result = bunnyEarsRecursive(bunnies-1);
-    return 2 + result;
-  }
-  return bunnyEarsRecursive(bunnies)
-}
-
-solutions.fibonacci = function (n) {
-  if (n < 2) {
-      return n;;
-  };
-    return fibonacci(n-1) + fibonacci(n-2)
-}
-
-solutions.bunnyEars2 = function (bunnies) {
-
- if (bunnies <= 0){
-   return 0;
- }
-
- if (bunnies%2 == 0){
-   return 3 + bunnyEars2(bunnies - 1);
- } else {
-   return 2 + bunnyEars2(bunnies - 1);
- }
-}
-
-solutions.triangle = function (rows) {
-
-  if (rows <= 0){
-    return 0;
-  }
-  return rows + triangle(rows-1)
-}
-
-solutions.sumDigits1 = function (n) {
-    if(n <= 0){
-       return 0;
-    }
-    return n % 10 + sumDigits1(Math.floor(n/10));
-}
-
-solutions.count7 = function (n) {
-    if(n <= 0)
-        return 0;
-
-    if(n % 10 == 7)
-        return 1 + count7(Math.floor(n/10));
-
-    return count7(Math.floor(n/10));
-}
-
-solutions.count8 = function (n) {
-    if(n == 0){
-      return 0;
-    }
-
-    if(n % 10 == 8) {
-        if(parseInt(n/10) % 10 == 8) {
-           return 2 + count8(parseInt(n/10));
-        } else {
-          return 1 + count8(parseInt(n/10));
-        }
-    }
-    return count8(parseInt(n/10));
-}
-
-solutions.powerN = function (base, n) {
-  if (n == 1){
-    return base;
-  }
-  return base * powerN(base, n-1);
-}
-
-solutions.countX = function (str) {
-    if(str.length == 0)
-        return 0;
-
-    if(str.charAt(0) == 'x'){
-      return 1 + countX(str.substring(1));
-    } else {
-      return countX(str.substring(1));
-    }
-}
-
-solutions.countHi = function (str) {
-  if (str.length == 0){
-    return 0;
-  }
-
-  if (str.substring(0, 2) == "hi"){
-    return 1 + countHi(str.substring(2))
-  }
-  return countHi(str.substring(1))
-}
-
-solutions.changeXY = function (str) {
-  if (str.length == 0){
-    return "";
-  }
-
-  if (str.charAt(0) == "x"){
-    return "y" + changeXY(str.substring(1))
-  }
-  return str.charAt(0) + changeXY(str.substring(1))
-}
-
-solutions.changePi = function (str) {
-
-  if (str.length <= 0){
-    return "";
-  }
-
-  if (str.substring(0, 2) == "pi"){
-    return 3.14 + changePi(str.substring(2))
-  } else {
-    return str.charAt(0) + changePi(str.substring(1))
-  }
-}
-
-solutions.noX = function (str) {
-  if (str.length == 0){
-    return "";
-  }
-
-  if (str.charAt(0) == "x"){
-    return noX(str.substring(1));
-  } else {
-    return str.charAt(0) + noX(str.substring(1));
-  }
-}
-
-solutions.array6 = function (nums, i) {
-  if (i > nums.length){
-    return false;
-  }
-
-  if (nums[i] == 6){
-    return true;
-  } else {
-    return array6(nums, i+1)
-  }
-}
-
-solutions.array11 = function array11(nums, i) {
-  if (i >= nums.length){
-    return 0;
-  }
-
-  if (nums[i] == 11){
-    return 1 + array11(nums, i+1);
-  } else {
-    return array11(nums, i+1);
-  }
-}
-
-solutions.array220 = function (nums, i) {
-
-  if (i == nums.length-1 || nums.length == 0){
-    return false;
-  }
-
-  if (nums[i] * 10 == nums[i+1]){
-    return true;
-  }
-  return array220(nums, i+1);
-}
-
-solutions.allStar = function (str) {
-
-  if (str.length <= 1){
-    return str;
-  }
-
-  return  str.charAt(0) + "*" + allStar(str.substring(1));
-}
-
-solutions.pairStar = function (str) {
-
-  if (str.length <= 1){
-    return str.charAt(0);
-  }
-
-  if (str.charAt(0) == str.charAt(1)){
-    return str.charAt(0) + "*" + pairStar(str.substring(1));
-  }
-  return str.charAt(0) + pairStar(str.substring(1));
-}
-
-solutions.endX = function (str) {
-    if(str.length <= 1) {
-      return str;
-    }
-
-   if (str.charAt(0) == "x"){
-     return endX(str.substring(1)) + "x";
-   }
-   return str.charAt(0) + endX(str.substring(1));
-}
-
-solutions.countPairs = function (str) {
-
-  if (str.length <= 2){
-    return 0;
-  }
-
-  if (str.charAt(0) == str.charAt(2)){
-    return 1 + countPairs(str.substring(1));
-  }
-  return countPairs(str.substring(1));
-}
-
-solutions.countAbc = function (str){
-
-  if (str.length <= 2) {
-    return 0;
-  }
-
-  if (str.substring(0, 3) === "abc" || str.substring(0, 3) === "aba"){
-    return 1 + countAbc(str.substring(2));
-  }
-  return countAbc(str.substring(1))
-}
-
-solutions.count11 = function (str) {
-
-  if (str.length <= 1){
-    return 0;
-  }
-
-  if (str.substring(0, 2) == "11"){
-    return 1 + count11(str.substring(2))
-  }
-  return count11(str.substring(1));
-}
-
-solutions.stringClean = function (str) {
-  if (str.length == 1){
-    return str;
-  }
-
-  if (str.charAt(0) == str.charAt(1)){
-    return stringClean(str.substring(1));
-  }
-  return str.charAt(0) + stringClean(str.substring(1));
-}
-
-solutions.countHi2 = function (str){
-  if (str.length <= 1){
-    return 0;
-  }
-
-  if (str.substring(0, 3) == "xhi"){
-    return countHi2(str.substring(3))
-  }
-
-  if (str.substring(0, 2) == "hi"){
-     return 1 + countHi2(str.substring(2))
-  }
-
-   return countHi2(str.substring(1))
-}
-
-solutions.starBit = function (str) {
-  if (str.charAt(0) == "-" && str.charAt(str.length - 1) == "*"){
-    return str;
-  }
-
-  if (str.charAt(str.length-1) == "*"){
-    return starBit(str.substring(1));
-  }
-
-  if (str.charAt(0) == "-"){
-    return starBit(str.substring(0, str.length-1));
-  }
-
-  return starBit(str.substring(1, str.length - 1));
-}
-
-solutions.nestParen = function (str) {
-    if(str.length == 0)
-        return true;
-
-    if(str.charAt(0) == '(' && str.charAt(str.length-1) == ')'){
-        return nestParen(str.substring(1, str.length-1));
-    }
-
-    return false;
-}
-
-solutions.strCount = function (str, sub) {
-  if (str.length <= 0){
-    return 0;
-  }
-
-  if (str.substring(0, sub.length) == sub){
-    return 1 + strCount(str.substring(sub.length), sub)
-  }
-  return strCount(str.substring(1), sub);
-}
-
-solutions.strCopies = function (str, sub, n) {
-  if (n == 0){
-    return true;
-  }
-
-  if (str.length < sub.length){
-    return false;
-  }
-
-  if (str.substring(0, sub.length) == sub){
-    return strCopies(str.substring(1), sub, n-1)
-  }
-  return strCopies(str.substring(1), sub, n)
-}
-
-solutions.strDist = function (str, sub) {
-
-  if (str.length <= sub.length){
-        if (str == sub){
-        return str.length;
-      } else {
-          return 0;
-      }
-  }
-
-  if (str.substring(0, sub.length) == sub && str.substring(str.length-sub.length) == sub){
-    return str.length;
-  }
-
-  if (str.substring(0, sub.length) == sub){
-    return strDist(str.substring(0, str.length-1), sub);
-  }
-
-   if (str.substring(str.length-sub.length) == sub){
-    return strDist(str.substring(1), sub);
-  }
-
-  return strDist(str.substring(1, str.length-1), sub)
-}
+];
