+++ src/utility/deParam.js
@@ -1,0 +1,8 @@
+/** takes url parameter and transforms into object **/
+function deParam(urlStr) {
+  const subbie = urlStr.substring(1);
+  const x = subbie.split('&').map(s => s.split('='));
+  return x.reduce((accum, arr) => { accum[arr[0]] = arr[1]; return accum; }, {});
+}
+
+module.exports = deParam;
