+++ src/utility/tableHeader.js
@@ -1,0 +1,9 @@
+function tableHeader() {
+    return `<tr>
+                <th>Test → Expected</th>
+                <th>Yours</th>
+                <th colspan="2">Result</th>
+            </tr>`;
+}
+
+module.exports = tableHeader;
